Initial commit to change core structure

+++ secbox/adb_dump_app.py
@@ -7,16 +7,18 @@ ADB = None
 def dump_apk(app_name, target_dir):
     if len(target_dir) == 0:
         print('[!] Invalid dir to save apk file!')
-        return
+        return False
 
     ADB = check_adb()
     os.system(ADB + ' shell pm path %s >> /tmp/%s' % (app_name, app_name))
     with open('/tmp/%s' % app_name, 'r') as f:
         source=f.read().split('\n')[0].replace('package:','').replace('\n','')
         dump_command=ADB + ' pull %s %s/%s.apk' % (source, target_dir, app_name)
-        os.system(dump_command)
-        os.system('rm -rf /tmp/%s' % app_name) 
-        print('\n[>] %s.apk extracted to %s' % (app_name, target_dir))
+        finished = os.system(dump_command) == 0
+        if finished:
+            os.system('rm -rf /tmp/%s' % app_name) 
+            print('\n[>] %s.apk extracted to %s' % (app_name, target_dir))
+        return finished
 
 def list_packages(app_name):
     ADB = check_adb()
